requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.4.0
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.2
chromadb==1.4.1
click==8.1.8
coloredlogs==15.0.1
cryptography==46.0.4
distro==1.9.0
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
feedparser==6.0.12
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.2
flatbuffers==25.12.19
fsspec==2025.5.1
ftfy==6.3.1
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.47.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==3.2.0
hf-xet==1.1.5
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==35.0.0
llvmlite==0.43.0
lxml==6.0.0
macholib==1.16.3
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
modulegraph==0.19.6
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.44.0
networkx==3.2.1
notion-client==2.7.0
numba==0.60.0
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
open_clip_torch==2.32.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.2.1
pinterest-crawler==0.2.0
playwright==1.52.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==5.29.5
py2app==0.28.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
pypdfium2==5.3.0
PyPika==0.51.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.3.2
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.33.0
sentence-transformers==5.1.2
sgmllib3k==1.0.0
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
streamlit==1.46.0
streamlit-tree-select==0.0.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.2
toml==0.10.2
tomli==2.4.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.21.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.5
wsproto==1.2.0
zipp==3.23.0
